Dockerfile
As I am not fond of golang, I don't want to install it on my machine.
But I am used to docker.

This commit adds a Dockerfile that is suitable for an automated build on
dockerhub.

Since the automated build is not configured yet, I added the steps to build the
image locally in the README.

Closes #14